backend/requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.30.5
python-dotenv==1.0.1
pydantic==2.8.2
supabase==2.6.0
openai==1.43.0
langchain==0.2.14
langchain-openai==0.1.22
langchain-community==0.2.12
pgvector==0.2.5
psycopg[binary]==3.2.1

